Broke out and improved testing
