+++ block-node/app/src/testFixtures/java/org/hiero/block/node/app/fixtures/plugintest/PluginTestBase.java
@@ -205,7 +205,8 @@ public abstract class PluginTestBase<
      * Teardown after each.
      */
     @AfterEach
-    public void tearDown() throws IOException {
+    public void tearDown() {
+        if (metricsProvider != null) metricsProvider.stop();
         testThreadPoolManager.shutdownNow();
         if (metricsRegistry != null) metricsRegistry.close();
     }
